requirements.txt
APScheduler==3.6.3
certifi==2021.5.30
chardet==4.0.0
geojson==2.5.0
idna==2.10
pyowm==3.2.0
PySocks==1.7.1
python-dotenv==0.17.1
python-telegram-bot==13.5
pytz==2021.1
requests==2.25.1
six==1.16.0
tornado==6.1
tzlocal==2.1
urllib3==1.26.5